+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 9.0.0-dev
 - BC break: methods startUp/shutDown/setUp/tearDown are not automatically run
 - BC break: removed Job::$onAfterExecute
 - BC break: removed prefix I from interfaces' names
+- argument --filterExceptFolders in mytester.php now applies also to .phpt files
 
 Version 8.2.1
 - data set is shown for skipped jobs or jobs passed with warnings